PackageSystemGenerator for collecting common process arguemnt function
