+++ bloom/generators/common.py
@@ -36,9 +36,15 @@ import pkg_resources
 import sys
 import traceback
 
+from bloom.git import inbranch
+from bloom.git import get_branches
+
 from bloom.logging import debug
 from bloom.logging import error
 from bloom.logging import info
+from bloom.logging import warning
+
+from bloom.packages import get_package_data
 
 from bloom.rosdistro_api import get_distribution_type
 
@@ -51,9 +57,15 @@ try:
     from rosdep2.catkin_support import get_catkin_view
     from rosdep2.lookup import ResolutionError
     import rosdep2.catkin_support
-except ImportError as err:
+except ImportError:
     debug(traceback.format_exc())
     error("rosdep was not detected, please install it.", exit=True)
+
+try:
+    import rosdistro
+except ImportError:
+    debug(traceback.format_exc())
+    error("rosdistro was not detected, please install it.", exit=True)
 
 BLOOM_GROUP = 'bloom.generators'
 DEFAULT_ROS_DISTRO = 'indigo'
@@ -206,6 +218,28 @@ def resolve_dependencies(
             resolved_key = fallback_resolver(key, peer_packages, os_name, os_version, ros_distro)
         resolved_keys[key] = resolved_key
     return resolved_keys
+
+
+def match_branches_with_prefix(prefix, get_branches, prune=False):
+    debug("match_branches_with_prefix(" + str(prefix) + ", " +
+          str(get_branches()) + ")")
+    branches = []
+    # Match branches
+    existing_branches = get_branches()
+    for branch in existing_branches:
+        if branch.startswith('remotes/origin/'):
+            branch = branch.split('/', 2)[-1]
+        if branch.startswith(prefix):
+            branches.append(branch)
+    branches = list(set(branches))
+    if prune:
+        # Prune listed branches by packages in latest upstream
+        with inbranch('upstream'):
+            pkg_names, version, pkgs_dict = get_package_data('upstream')
+            for branch in branches:
+                if branch.split(prefix)[-1].strip('/') not in pkg_names:
+                    branches.remove(branch)
+    return branches
 
 
 class GeneratorError(Exception):
@@ -363,3 +397,89 @@ class BloomGenerator(object):
         :returns: return code, return 0 or None for OK, anythign else on error
         """
         return 0
+
+
+class PackageSystemGenerator(BloomGenerator):
+    package_system = 'none'
+
+    def prepare_arguments(self, parser):
+        # The common command line arguments for every package system
+        add = parser.add_argument
+        add('-i', '--inc', help="increment number", default='0')
+        add('-p', '--prefix', required=True,
+            help="branch prefix to match, and from which create packages"
+                 " hint: if you want to match 'release/foo' use 'release'")
+        add('-a', '--match-all', default=False, action="store_true",
+            help="match all branches with the given prefix, "
+                 "even if not in current upstream")
+        add('--distros', nargs='+', required=False, default=[],
+            help='A list of os distros to generate for certain package system')
+        add('--install-prefix', default=None,
+            help="overrides the default installation prefix (/usr)")
+
+    def get_package_from_branch(self, branch):
+        with inbranch(branch):
+            try:
+                package_data = get_package_data(branch)
+            except SystemExit:
+                return None
+            if type(package_data) not in [list, tuple]:
+                # It is a ret code
+                self.exit(package_data)
+        names, version, packages = package_data
+        if type(names) is list and len(names) > 1:
+            self.exit(
+                "{0} generator does not support generating "
+                "from branches with multiple packages in them, use "
+                "the release generator first to split packages into "
+                "individual branches."
+                .format(self.package_system))
+        if type(packages) is dict:
+            return list(packages.values())[0]
+
+    def get_default_distros(self):
+        index = rosdistro.get_index(rosdistro.get_index_url())
+        distribution_file = rosdistro.get_distribution_file(index, self.rosdistro)
+        if self.os_name not in distribution_file.release_platforms:
+            if hasattr(self, "os_not_required") and self.os_not_required:
+                warning("No platforms defined for os '{0}' in release file for the "
+                        "'{1}' distro. This os was not required; continuing without error."
+                        .format(self.os_name, self.rosdistro))
+                sys.exit(0)
+            error("No platforms defined for os '{0}' in release file for the '{1}' distro."
+                  .format(self.os_name, self.rosdistro), exit=True)
+        self.distros = distribution_file.release_platforms[self.os_name]
+
+    def handle_arguments(self, args):
+        self.interactive = args.interactive
+        self.inc = args.inc
+        self.os_name = args.os_name
+        self.distros = args.distros
+        if self.distros in [None, []]:
+            self.get_default_distros()
+        self.install_prefix = args.install_prefix
+        if args.install_prefix is None:
+            self.install_prefix = self.default_install_prefix
+        self.prefix = args.prefix
+        self.branches = match_branches_with_prefix(self.prefix, get_branches, prune=not args.match_all)
+        if len(self.branches) == 0:
+            error(
+                "No packages found, check your --prefix or --src arguments.",
+                exit=True
+            )
+        self.packages = {}
+        self.tag_names = {}
+        self.names = []
+        self.branch_args = []
+        self.package_system_branches = []
+        for branch in self.branches:
+            package = self.get_package_from_branch(branch)
+            if package is None:
+                # This is an ignored package
+                continue
+            self.packages[package.name] = package
+            self.names.append(package.name)
+            args = self.generate_branching_arguments(package, branch)
+            # First branch is package_system/[<rosdistro>/]<package>
+            self.package_system_branches.append(args[0][0])
+            self.branch_args.extend(args)
